+++ api-testcases/test-logging.c
@@ -13,13 +13,16 @@ TIMESTAMP FTL C: clamped fatal: 123
 TIMESTAMP INF C: long long long long long long long long long long long long long long long long long long long long long long long long long long long long long long long long long long long long long long long long long long long long long long long long long long long long long long long long long long long long long long long long long long long long long long long long long long long long long long long long long long long long long long long long long long long long long long long long long long long long long long long long long long long long long long long long long long long long long long long long long long long long long long long long long long long long long long long long long long long long long long long long long long long long long long long long long long long long long long long long long long long long long long long long long long long long long long long long long long long long long long long long long long long long long long long long long long long long long long long long and formatted: 123
 final top: 0
 ==> rc=0, result='undefined'
+*** test_2 (duk_safe_call)
+TIMESTAMP FTL C: fatal error: 123 quux
+==> rc=0, result='undefined'
 ===*/
 
 #define  CHARS_100 \
 	"long long long long long long long long long long " \
 	"long long long long long long long long long long "
 
-int test_1(duk_context *ctx) {
+static duk_ret_t test_1(duk_context *ctx) {
 	/* Force log level to output all logs. */
 	duk_eval_string(ctx, "Duktape.Logger.clog.l = 0;");
 	duk_pop(ctx);
@@ -60,6 +63,23 @@ int test_1(duk_context *ctx) {
 	return 0;
 }
 
+/* Vararg */
+static void my_log_write(duk_context *ctx, duk_int_t level, const char *fmt, ...) {
+	va_list ap;
+
+	va_start(ap, fmt);
+	duk_log_va(ctx, level, fmt, ap);
+	va_end(ap);
+}
+
+static duk_ret_t test_2(duk_context *ctx) {
+	/* Rely on the log "censoring" set up in test_1(). */
+
+	my_log_write(ctx, DUK_LOG_FATAL, "fatal error: %d %s", 123, "quux");
+	return 0;
+}
+
 void test(duk_context *ctx) {
 	TEST_SAFE_CALL(test_1);
+	TEST_SAFE_CALL(test_2);
 }